drivers/exec: SIGINT unavailable on windows
