+++ drivers/exec/driver_test.go
@@ -123,66 +123,6 @@ func TestExecDriver_StartWait(t *testing.T) {
 	require.NoError(harness.DestroyTask(task.ID, true))
 }
 
-func TestExecDriver_StartWaitStop(t *testing.T) {
-	t.Parallel()
-	require := require.New(t)
-	ctestutils.ExecCompatible(t)
-
-	d := NewExecDriver(testlog.HCLogger(t))
-	harness := dtestutil.NewDriverHarness(t, d)
-	task := &drivers.TaskConfig{
-		ID:        uuid.Generate(),
-		Name:      "test",
-		Resources: testResources,
-	}
-
-	taskConfig := map[string]interface{}{
-		"command": "/bin/sleep",
-		"args":    []string{"600"},
-	}
-	encodeDriverHelper(require, task, taskConfig)
-
-	cleanup := harness.MkAllocDir(task, false)
-	defer cleanup()
-
-	handle, _, err := harness.StartTask(task)
-	require.NoError(err)
-
-	ch, err := harness.WaitTask(context.Background(), handle.Config.ID)
-	require.NoError(err)
-
-	require.NoError(harness.WaitUntilStarted(task.ID, 1*time.Second))
-
-	go func() {
-		harness.StopTask(task.ID, 2*time.Second, "SIGINT")
-	}()
-
-	select {
-	case result := <-ch:
-		require.Equal(int(unix.SIGINT), result.Signal)
-	case <-time.After(10 * time.Second):
-		require.Fail("timeout waiting for task to shutdown")
-	}
-
-	// Ensure that the task is marked as dead, but account
-	// for WaitTask() closing channel before internal state is updated
-	testutil.WaitForResult(func() (bool, error) {
-		status, err := harness.InspectTask(task.ID)
-		if err != nil {
-			return false, fmt.Errorf("inspecting task failed: %v", err)
-		}
-		if status.State != drivers.TaskStateExited {
-			return false, fmt.Errorf("task hasn't exited yet; status: %v", status.State)
-		}
-
-		return true, nil
-	}, func(err error) {
-		require.NoError(err)
-	})
-
-	require.NoError(harness.DestroyTask(task.ID, true))
-}
-
 func TestExecDriver_StartWaitStopKill(t *testing.T) {
 	t.Parallel()
 	require := require.New(t)
